requirements.txt
certifi==2020.6.20
chardet==3.0.4
idna==2.10
psycopg2-binary==2.8.5
python-dotenv==0.14.0
redis==3.5.3
requests==2.24.0
urllib3==1.25.9
